# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/moderation/ban/{communityId}/{userId}": {
      "post": {
        "operationId": "ModerationController_banUser",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationBanUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      },
      "delete": {
        "operationId": "ModerationController_unbanUser",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbanUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/bans/{communityId}": {
      "get": {
        "operationId": "ModerationController_getBanList",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommunityBanDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/kick/{communityId}/{userId}": {
      "post": {
        "operationId": "ModerationController_kickUser",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KickUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/timeout/{communityId}/{userId}": {
      "post": {
        "operationId": "ModerationController_timeoutUser",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeoutUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      },
      "delete": {
        "operationId": "ModerationController_removeTimeout",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTimeoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/timeouts/{communityId}": {
      "get": {
        "operationId": "ModerationController_getTimeoutList",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommunityTimeoutDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/timeout-status/{communityId}/{userId}": {
      "get": {
        "operationId": "ModerationController_getTimeoutStatus",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeoutStatusResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/pin/{messageId}": {
      "post": {
        "operationId": "ModerationController_pinMessage",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      },
      "delete": {
        "operationId": "ModerationController_unpinMessage",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnpinMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/pins/{channelId}": {
      "get": {
        "operationId": "ModerationController_getPinnedMessages",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PinnedMessageDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/message/{messageId}": {
      "delete": {
        "operationId": "ModerationController_deleteMessageAsMod",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMessageAsModDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/moderation/logs/{communityId}": {
      "get": {
        "operationId": "ModerationController_getModerationLogs",
        "parameters": [
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationLogsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kraken API",
    "description": "Kraken Chat Application API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SpanDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "PLAINTEXT",
              "USER_MENTION",
              "SPECIAL_MENTION",
              "COMMUNITY_MENTION",
              "ALIAS_MENTION"
            ]
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "specialKind": {
            "type": "string",
            "nullable": true
          },
          "communityId": {
            "type": "string",
            "nullable": true
          },
          "aliasId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "type",
          "text",
          "userId",
          "specialKind",
          "communityId",
          "aliasId"
        ]
      },
...
      "ReactionDto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string"
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "emoji",
          "userIds"
        ]
      },
...
      "SuccessMessageDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "UnbanUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "ModerationUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "USER"
            ]
          },
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "lastSeen": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "role",
          "id",
          "username",
          "avatarUrl",
          "bannerUrl",
          "lastSeen",
          "displayName",
          "bio",
          "status"
        ]
      },
      "CommunityBanDto": {
        "type": "object",
        "properties": {
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ModerationUserDto"
              }
            ]
          },
          "moderator": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ModerationUserDto"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "moderatorId": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "user",
          "moderator",
          "id",
          "communityId",
          "userId",
          "moderatorId",
          "reason",
          "createdAt",
          "expiresAt",
          "active"
        ]
      },
      "KickUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "TimeoutUserDto": {
        "type": "object",
        "properties": {
          "durationSeconds": {
            "type": "number",
            "minimum": 60,
            "maximum": 2419200
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "durationSeconds"
        ]
      },
      "RemoveTimeoutDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "CommunityTimeoutDto": {
        "type": "object",
        "properties": {
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ModerationUserDto"
              }
            ]
          },
          "moderator": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ModerationUserDto"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "moderatorId": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "user",
          "moderator",
          "id",
          "communityId",
          "userId",
          "moderatorId",
          "reason",
          "createdAt",
          "expiresAt"
        ]
      },
      "TimeoutStatusResponseDto": {
        "type": "object",
        "properties": {
          "isTimedOut": {
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "isTimedOut"
        ]
      },
      "PinMessageDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "UnpinMessageDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "PinnedMessageAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "avatarUrl"
        ]
      },
      "PinnedMessageAttachmentDto": {
        "type": "object",
        "properties": {
          "fileType": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "AUDIO",
              "DOCUMENT",
              "OTHER"
            ]
          },
          "id": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          }
        },
        "required": [
          "fileType",
          "id",
          "filename",
          "mimeType",
          "size"
        ]
      },
      "PinnedMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "nullable": true
          },
          "directMessageGroupId": {
            "type": "string",
            "nullable": true
          },
          "authorId": {
            "type": "string"
          },
          "spans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpanDto"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReactionDto"
            }
          },
          "sentAt": {
            "format": "date-time",
            "type": "string"
          },
          "editedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "pinned": {
            "type": "boolean"
          },
          "pinnedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "pinnedBy": {
            "type": "string",
            "nullable": true
          },
          "replyCount": {
            "type": "number"
          },
          "lastReplyAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "searchText": {
            "type": "string",
            "nullable": true
          },
          "pendingAttachments": {
            "type": "number",
            "nullable": true
          },
          "deletedBy": {
            "type": "string",
            "nullable": true
          },
          "deletedByReason": {
            "type": "string",
            "nullable": true
          },
          "parentMessageId": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PinnedMessageAuthorDto"
              }
            ]
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PinnedMessageAttachmentDto"
            }
          }
        },
        "required": [
          "id",
          "channelId",
          "directMessageGroupId",
          "authorId",
          "spans",
          "reactions",
          "sentAt",
          "editedAt",
          "deletedAt",
          "pinned",
          "pinnedAt",
          "pinnedBy",
          "replyCount",
          "lastReplyAt",
          "searchText",
          "pendingAttachments",
          "deletedBy",
          "deletedByReason",
          "parentMessageId",
          "author",
          "attachments"
        ]
      },
      "DeleteMessageAsModDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "ModerationLogsResponseDto": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModerationLogDto"
            }
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "logs",
          "total"
        ]
      },
...
      "ModerationBanUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        }
      },
...
    }
  }
}